Unify Ideal/Robust Soliton and Encoder
