+++ src/decoder.rs
@@ -1,7 +1,7 @@
 use crate::{
     block::Block,
     droplet::{Droplet, RxDroplet},
-    robust_encoder::get_sample_from_rng_by_seed,
+    encoder::get_sample_from_rng_by_seed,
     types::{CatchResult, DropType},
 };
 use rand::distributions::Uniform;
@@ -36,7 +36,6 @@ impl Decoder {
     /// extern crate fountaincode;
     ///
     /// fn main() {
-    ///     use self::fountaincode::ideal_encoder::IdealEncoder;
     ///     use self::fountaincode::decoder::Decoder;
     ///     use self::fountaincode::encoder::Encoder;
     ///     use self::fountaincode::types::*;
@@ -48,11 +47,11 @@ impl Decoder {
     ///     let to_compare = buf.clone();
     ///     let length = buf.len();
     ///
-    ///     let mut enc = IdealEncoder::new(buf, 64, EncoderType::Random);
+    ///     let mut enc = Encoder::ideal(buf, 64, EncoderType::Random);
     ///     let mut dec = Decoder::new(length, 64);
     ///
     ///     loop {
-    ///         let drop = enc.next();
+    ///         let drop = enc.drop();
     ///         match dec.catch(drop) {
     ///             CatchResult::Missing(stats) => {
     ///                 println!("Missing blocks {:?}", stats);
