+++ src/lib.rs
@@ -4,9 +4,5 @@ pub mod block;
 pub mod decoder;
 pub mod droplet;
 pub mod encoder;
-pub mod ideal_encoder;
-pub mod ideal_soliton;
-pub mod robust_encoder;
-pub mod robust_soliton;
 pub mod soliton;
 pub mod types;
